refactor(durability): split near-limit modules
