+++ crates/convergio-durability/src/store/crdt.rs
@@ -1,6 +1,7 @@
 //! CRDT actor and operation store.
 
 use crate::error::{DurabilityError, Result};
+use crate::store::crdt_rows::{CrdtActorRow, CrdtOpRow};
 use chrono::{DateTime, Utc};
 use convergio_db::Pool;
 use serde::{Deserialize, Serialize};
@@ -218,30 +219,6 @@ const SELECT_OP: &str = "SELECT actor_id, counter, entity_type, entity_id, field
      op_kind, value, hlc, created_at FROM crdt_ops \
      WHERE actor_id = ? AND counter = ? LIMIT 1";
 
-#[derive(sqlx::FromRow)]
-struct CrdtActorRow {
-    actor_id: String,
-    kind: String,
-    display_name: Option<String>,
-    is_local: i64,
-    created_at: String,
-    last_seen_at: String,
-}
-
-#[derive(sqlx::FromRow)]
-struct CrdtOpRow {
-    actor_id: String,
-    counter: i64,
-    entity_type: String,
-    entity_id: String,
-    field_name: String,
-    crdt_type: String,
-    op_kind: String,
-    value: String,
-    hlc: String,
-    created_at: String,
-}
-
 impl CrdtOp {
     fn same_identity_payload(&self, other: &NewCrdtOp) -> bool {
         self.entity_type == other.entity_type
@@ -252,45 +229,4 @@ impl CrdtOp {
             && self.value == other.value
             && self.hlc == other.hlc
     }
-}
-
-impl TryFrom<CrdtActorRow> for CrdtActor {
-    type Error = DurabilityError;
-    fn try_from(row: CrdtActorRow) -> Result<Self> {
-        Ok(Self {
-            actor_id: row.actor_id,
-            kind: row.kind,
-            display_name: row.display_name,
-            is_local: row.is_local == 1,
-            created_at: parse_ts(&row.created_at)?,
-            last_seen_at: parse_ts(&row.last_seen_at)?,
-        })
-    }
-}
-
-impl TryFrom<CrdtOpRow> for CrdtOp {
-    type Error = DurabilityError;
-    fn try_from(row: CrdtOpRow) -> Result<Self> {
-        Ok(Self {
-            actor_id: row.actor_id,
-            counter: row.counter,
-            entity_type: row.entity_type,
-            entity_id: row.entity_id,
-            field_name: row.field_name,
-            crdt_type: row.crdt_type,
-            op_kind: row.op_kind,
-            value: serde_json::from_str(&row.value)?,
-            hlc: row.hlc,
-            created_at: parse_ts(&row.created_at)?,
-        })
-    }
-}
-
-fn parse_ts(s: &str) -> Result<DateTime<Utc>> {
-    DateTime::parse_from_rfc3339(s)
-        .map(|d| d.with_timezone(&Utc))
-        .map_err(|_| DurabilityError::NotFound {
-            entity: "timestamp",
-            id: s.to_string(),
-        })
 }
